Add state & quadtree

+++ MegaManX3/Textures.h
@@ -10,22 +10,21 @@ using namespace std;
 struct STexture
 {
 	LPDIRECT3DTEXTURE9 texture;
-	float width;
-	float height;
+	UINT width;
+	UINT height;
 };
 
 class CTextures
 {
 	static CTextures * __instance;
-
-	unordered_map<int, STexture*> textures;
+	unordered_map<unsigned int, STexture*> textures;
 
 public:
-	void add(int id, LPCWSTR filePath, float width = 0, float height = 0, D3DCOLOR transparentColor = BLACK(0));
-	LPDIRECT3DTEXTURE9 getTexture(int id);
-	void getSize(int id, int &width, int &height);
-	STexture* getSTexture(int id);
-	void deleteAt(int id);
+	void add(UINT id, LPCWSTR filePath, UINT width = 0, UINT height = 0, D3DCOLOR transparentColor = BLACK(0));
+	LPDIRECT3DTEXTURE9 getTexture(UINT id);
+	void getSize(UINT id, UINT &width, UINT &height);
+	STexture* getSTexture(UINT id);
+	void deleteAt(UINT id);
 	void clear();
 
 	static CTextures * getInstance();
